Move members builder functions to ExplorePage
